move TypeInfo implementation into cpp
