+++ be/src/storage/decimal_type_info.h
@@ -19,9 +19,7 @@ public:
     using CppType = typename CppTypeTraits<TYPE>::CppType;
     using Datum = vectorized::Datum;
     DecimalTypeInfo(int precision, int scale)
-            : _delegate(ScalarTypeInfoResolver::instance()->get_scalar_type_info(DelegateType<TYPE>)),
-              _precision(precision),
-              _scale(scale) {
+            : _delegate(get_scalar_type_info(DelegateType<TYPE>)), _precision(precision), _scale(scale) {
         static_assert(!ft_is_invalid<DelegateType<TYPE>>);
     }
